src/requirements.txt
python-dateutil==2.8.2
pytz==2024.1
google-auth==2.36.0
langchain-google-vertexai==2.0.9
langchain-google-genai==2.0.7
langchain==0.3.9
langchain-core==0.3.29
qdrant-client==1.12.2
streamlit==1.41.1
pydantic==2.9.2
python-dotenv==1.0.1
pathlib
protobuf>=3.20,<3.21
google-cloud-aiplatform<1.72.0
langchain-community==0.3.8
langchain_huggingface==0.1.2
langchain_qdrant==0.2.0
langgraph==0.2.60
langdetect==1.0.9
langchain-groq==0.2.3
ipython
rank-bm25==0.2.2 